+++ packages/tce-boot/CHANGELOG.md
@@ -1,5 +1,147 @@
 # @tailor-cms/tce-boot
 
+## 0.6.27
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.27
+
+## 0.6.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.26
+  - @tailor-cms/tce-server-runtime@0.5.5
+
+## 0.6.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.25
+
+## 0.6.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.24
+  - @tailor-cms/tce-server-runtime@0.5.4
+
+## 0.6.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.23
+
+## 0.6.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.22
+
+## 0.6.21
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.21
+
+## 0.6.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.20
+  - @tailor-cms/tce-display-runtime@0.6.7
+
+## 0.6.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.19
+
+## 0.6.18
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.18
+
+## 0.6.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.17
+  - @tailor-cms/tce-display-runtime@0.6.6
+
+## 0.6.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.16
+
+## 0.6.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.15
+  - @tailor-cms/tce-display-runtime@0.6.5
+
+## 0.6.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.14
+  - @tailor-cms/tce-server-runtime@0.5.3
+
+## 0.6.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.13
+  - @tailor-cms/tce-server-runtime@0.5.2
+
+## 0.6.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.12
+
+## 0.6.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.11
+  - @tailor-cms/tce-display-runtime@0.6.4
+
+## 0.6.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.10
+  - @tailor-cms/tce-display-runtime@0.6.3
+
+## 0.6.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-next-runtime@0.6.9
+
 ## 0.6.8
 
 ### Patch Changes